pyproject.toml
[project]
name = "lambda-classification-summarization-recman-test-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-openai>=0.3.35",
    "pydantic>=2.12.0",
]

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.219.0",
    "black>=25.9.0",
    "constructs>=10.4.2",
    "flake8>=7.3.0",
    "pytest>=8.4.2",
]
